# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Host App API - Members-only platform for artists
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface MessageResponse {
  message: string;
}

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  pending: "pending",
  active: "active",
  suspended: "suspended",
} as const;

export interface User {
  id: number;
  email: string;
  displayName: string;
  pseudonym?: string | null;
  isAdmin: boolean;
  status: UserStatus;
  createdAt: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 2 */
  displayName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type ApplicationRequestPersonalityAnswers = { [key: string]: unknown };

export interface ApplicationRequest {
  /** @minimum 18 */
  age: number;
  nationality: string;
  profession: string;
  educationalBackground?: string;
  personalityAnswers?: ApplicationRequestPersonalityAnswers;
  skillTags?: string[];
  artistStatement: string;
}

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  submitted: "submitted",
  processing: "processing",
  assigned: "assigned",
} as const;

export interface Application {
  id: number;
  userId: number;
  cohortId?: number | null;
  applicationOrder?: number | null;
  age: number;
  nationality: string;
  profession: string;
  educationalBackground?: string | null;
  skillTags?: string[];
  artistStatement: string;
  status: ApplicationStatus;
  submittedAt: string;
}

export type ApplicationWithUser = Application & {
  userDisplayName: string;
  userEmail: string;
};

export type CohortStatusProperty =
  (typeof CohortStatusProperty)[keyof typeof CohortStatusProperty];

export const CohortStatusProperty = {
  open: "open",
  processing: "processing",
  active: "active",
  closed: "closed",
} as const;

export interface Cohort {
  id: number;
  cohortNumber: number;
  applicantCount: number;
  status: CohortStatusProperty;
  createdAt: string;
  lockedAt?: string | null;
}

export interface CohortStatus {
  cohortNumber: number;
  applicantCount: number;
  spotsRemaining: number;
  status: string;
}

export interface CohortProcessResult {
  cohortId: number;
  rolesAssigned: number;
  teamACount: number;
  teamBCount: number;
  leaderAssigned: boolean;
  peripheralCount: number;
}

export type UserRoleRoleType =
  (typeof UserRoleRoleType)[keyof typeof UserRoleRoleType];

export const UserRoleRoleType = {
  team_member: "team_member",
  leader: "leader",
  peripheral: "peripheral",
} as const;

export interface UserRole {
  roleType: UserRoleRoleType;
  teamName?: string | null;
  isHiddenRole: boolean;
  cohortNumber: number;
  statusLabel: string;
  assignedAt: string;
}

export type RoomRoomType = (typeof RoomRoomType)[keyof typeof RoomRoomType];

export const RoomRoomType = {
  team_a: "team_a",
  team_b: "team_b",
  leader: "leader",
  peripheral: "peripheral",
  admin_broadcast: "admin_broadcast",
} as const;

export interface Room {
  id: number;
  cohortId: number;
  roomType: RoomRoomType;
  visibilityRule: string;
  memberCount: number;
}

export interface Message {
  id: number;
  roomId: number;
  userId?: number | null;
  content: string;
  isSystemMessage: boolean;
  maskedSenderLabel?: string | null;
  createdAt: string;
}

export interface SendMessageRequest {
  /** @minLength 1 */
  content: string;
}

export interface Instruction {
  id: number;
  cohortId: number;
  targetRole?: string | null;
  content: string;
  releaseAt?: string | null;
  createdAt: string;
}

export interface CreateInstructionRequest {
  cohortId: number;
  targetRole?: string | null;
  targetUserId?: number | null;
  content: string;
  releaseAt?: string | null;
}

export interface AdminStats {
  totalUsers: number;
  totalApplications: number;
  totalCohorts: number;
  activeCohorts: number;
  openCohortApplicants: number;
  spotsToFill: number;
}

export type GetRoomMessagesParams = {
  limit?: number;
  offset?: number;
};