# lib/api-zod/src/generated/types/applicationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Host App API - Members-only platform for artists
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationRequestPersonalityAnswers } from "./applicationRequestPersonalityAnswers";

export interface ApplicationRequest {
  /** @minimum 18 */
  age: number;
  nationality: string;
  profession: string;
  educationalBackground?: string;
  personalityAnswers?: ApplicationRequestPersonalityAnswers;
  skillTags?: string[];
  artistStatement: string;
}